FINISH

+++ application/views/memo/index.php
@@ -2,13 +2,19 @@
 <div class="container-fluid">
 
   <!-- Page Heading -->
-  <h1 class="h3 mb-2 text-gray-800">Table Memo</h1>
-
+  <h1 class="h3 mb-2 text-gray-800">Table Memo Processed</h1>
+  <div class="row">
+    <div class="col-lg-12">
+      <?= $this->session->flashdata('message'); ?>
+    </div>
+  </div>
   <!-- DataTales Example -->
   <div class="card shadow mb-4">
-    <div class="card-header py-3">
-      <a href="<?= base_url('memo/tambah'); ?>" class="btn btn-primary btn-sm">Buat Memo</a>
-    </div>
+    <?php if ($user['user_role'] != 'Direktur'): ?>
+      <div class="card-header py-3">
+        <a href="<?= base_url('memo/tambah'); ?>" class="btn btn-primary btn-sm">Buat Memo</a>
+      </div>
+    <?php endif ?>
     <div class="card-body">
       <div class="table-responsive">
         <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
@@ -31,9 +37,9 @@
                 <?php if ($mm['mm_status'] == 1): ?>
                   <tr align="center">
                     <td><?= $no++ ?></td>
-                    <td><?= $mm['mm_no']; ?></td>
+                    <td><?= $mm['mm_id']; ?> / MI / PTHG</td>
                     <td><?= $mm['user_name']; ?><br>
-                      (<?= $mm['user_role']; ?>-<?= $mm['user_bagian']; ?>)
+                      (<?= $mm['user_role']; ?> <?= $mm['user_bagian']; ?>)
                     </td>
                     <td><?= $mm['mm_perihal']; ?></td>
                     <td><?= date('d F Y', $mm['mm_tgl_buat']); ?></td>
@@ -54,7 +60,7 @@
                         <?php if ($mm['mm_status'] != 3 ): ?>
                           <a href="<?= base_url(); ?>memo/edit/<?= $mm['mm_id']; ?>" class="btn btn-info btn-sm">Edit</a>
                         <?php endif ?>
-                        <a href="<?= base_url(); ?>memo/hapus/<?= $mm['mm_id']; ?>" class="btn btn-danger btn-sm">Hapus</a>
+                        <a onclick="return confirm('Yakin akan menghapus data?');" href="<?= base_url(); ?>memo/hapus/<?= $mm['mm_id']; ?>" class="btn btn-danger btn-sm">Hapus</a>
                       <?php endif ?>
                     </td>
                   </tr>
